fix: Separated Prizegiving and Shed settings forms
Separate update and delete actions for prizegivings and shed even though they load on the same page
