+++ app/Http/Controllers/Setting/EventsSettingController.php
@@ -37,14 +37,19 @@ class EventsSettingController extends Controller
                 $storagePath = Storage::disk('public')->put('/files', $request->file('shed_consentForm'));
                 $settings->shed_consentFormLink = $storagePath;
             }
+            if ($request->hasAny(['shed_dates', 'shed_location', 'shed_year', 'shed_embedLink', 'shed_isActive'])) {
 
-            $settings->shed_dates = $request->input('shed_dates');
-            $settings->shed_location = $request->input('shed_location');
-            $settings->shed_year = $request->input('shed_year');
-            $settings->shed_embedLink = $request->input('shed_embedLink');
-            $settings->shed_isActive = $request->input('shed_isActive');
-            $settings->prizegivings_year = $request->input('prizegivings_year');
-            $settings->prizegivings_isActive = $request->boolean('prizegivings_isActive');
+                $settings->shed_dates = $request->input('shed_dates');
+                $settings->shed_location = $request->input('shed_location');
+                $settings->shed_year = $request->input('shed_year');
+                $settings->shed_embedLink = $request->input('shed_embedLink');
+                $settings->shed_isActive = $request->input('shed_isActive');
+            }
+            if ($request->hasAny(['prizegivings_year', 'prizegivings_isActive'])) {
+
+                $settings->prizegivings_year = $request->input('prizegivings_year');
+                $settings->prizegivings_isActive = $request->boolean('prizegivings_isActive');
+            }
 
             $settings->save();
         } catch (\Exception $exception) {
@@ -54,21 +59,20 @@ class EventsSettingController extends Controller
         return redirect()->route('settings.events.index')->with('success', 'Events settings updated successfully');
     }
 
-    public function destroyFile(EventsSettings $settings)
+    public function destroyFile(Request $request, EventsSettings $settings)
     {
-        dd(Storage::disk('public')->exists(""));
-        if (Storage::disk('public')->exists($settings->prizegivings_scheduleFileLink)) {
-            Storage::disk('public')->delete($settings->prizegivings_scheduleFileLink);
+        if (isset($request->prizegivings_scheduleFileLink) && Storage::disk('public')->exists($request->prizegivings_scheduleFileLink)) {
+            Storage::disk('public')->delete($request->prizegivings_scheduleFileLink);
 
-            $settings->prizegivings_scheduleFileLink = "";
+            $settings->prizegivings_scheduleFileLink = null;
             $settings->save();
 
             return redirect()->route('settings.events.index')->with('success', 'Schedule file removed');
         }
-        if (Storage::disk('public')->exists($settings->shed_consentFormLink)) {
-            Storage::disk('public')->delete($settings->shed_consentFormLink);
+        if (isset($request->shed_consentFormLink) && Storage::disk('public')->exists($request->shed_consentFormLink)) {
+            Storage::disk('public')->delete($request->shed_consentFormLink);
 
-            $settings->shed_consentFormLink = "";
+            $settings->shed_consentFormLink = null;
             $settings->save();
 
             return redirect()->route('settings.events.index')->with('success', 'Consent form file removed');
